{"schema":"decodex/commit/1","summary":"Move scroll capture sample input","authority":"manual"}
